Split interpolation.core into two files
